use outlet
